State store the Redux
The overall concept is to use Redux to manage the authentication-related state of the application. The authSlice.js file defines the initial state, actions, and reducers related to authentication. The store.js file then creates a Redux store with this authentication slice, making it available for use in a React application. The state changes (like setting credentials or logging out) are handled through Redux actions, and the state is stored in the Redux store and, for persistence, in the local storage of the browser. This allows components in the React application to read from and dispatch actions to this centralized state store, making it easier to manage and share state across different parts of the application